Rahul/ifl 2399 move spendposttime to utils (#4859)
* add spend post time to utils

* using getSpendPostTime in notes combine
